Refactorized the parser
Moved some of the functionality in parser.c into new files.  There's
a lot more work to be done here but this is the first attempt to
sanitize the parser.
