+++ Makefile
@@ -8,6 +8,15 @@ COMMON =\
 	keywords.c \
 	buffer.c \
 	parser.c \
+	parser_array.c \
+	parser_call.c \
+	parser_compound.c \
+	parser_exp.c \
+	parser_math.c \
+	parser_error.c \
+	parser_graphics.c \
+	parser_input.c \
+	parser_rnd.c \
 	expression.c \
 	ir.c \
 	hash_table.c \
