+++ frontend/parser.c
@@ -25,39 +25,14 @@
 #include "expression.h"
 #include "globals.h"
 #include "parser.h"
+#include "parser_array.h"
+#include "parser_call.h"
+#include "parser_compound.h"
+#include "parser_error.h"
+#include "parser_exp.h"
 #include "type_if.h"
 #include "variable.h"
 
-struct subtilis_parser_param_t_ {
-	subtilis_type_t type;
-	size_t reg;
-	size_t nop;
-};
-
-typedef struct subtilis_parser_param_t_ subtilis_parser_param_t;
-
-static subtilis_exp_t *prv_call_1_arg_fn(subtilis_parser_t *p, const char *name,
-					 size_t reg,
-					 subtilis_ir_reg_type_t ptype,
-					 const subtilis_type_t *rtype,
-					 subtilis_error_t *err);
-static subtilis_exp_t *prv_priority1(subtilis_parser_t *p, subtilis_token_t *t,
-				     subtilis_error_t *err);
-static subtilis_exp_t *prv_call(subtilis_parser_t *p, subtilis_token_t *t,
-				subtilis_error_t *err);
-static subtilis_exp_t *prv_bracketed_2_int_args(subtilis_parser_t *p,
-						subtilis_token_t *t,
-						subtilis_op_instr_type_t itype,
-						subtilis_error_t *err);
-static void prv_add_fn_ret(subtilis_parser_t *p, const subtilis_type_t *fn_type,
-			   subtilis_error_t *err);
-static void prv_set_fn_retval(subtilis_parser_t *p, subtilis_exp_t *e,
-			      const subtilis_type_t *fn_type,
-			      subtilis_error_t *err);
-static size_t prv_var_bracketed_int_args_have_b(subtilis_parser_t *p,
-						subtilis_token_t *t,
-						subtilis_exp_t **e, size_t max,
-						subtilis_error_t *err);
 #define SUBTILIS_MAIN_FN "subtilis_main"
 
 subtilis_parser_t *subtilis_parser_new(subtilis_lexer_t *l,
@@ -160,1203 +135,6 @@ typedef void (*subtilis_keyword_fn)(subtilis_parser_t *p, subtilis_token_t *,
 
 static const subtilis_keyword_fn keyword_fns[];
 
-static subtilis_exp_t *prv_expression(subtilis_parser_t *p, subtilis_token_t *t,
-				      subtilis_error_t *err);
-
-static subtilis_exp_t *prv_bracketed_exp(subtilis_parser_t *p,
-					 subtilis_token_t *t,
-					 subtilis_error_t *err)
-{
-	const char *tbuf;
-	subtilis_exp_t *e;
-
-	e = prv_expression(p, t, err);
-	if (err->type != SUBTILIS_ERROR_OK)
-		return NULL;
-	tbuf = subtilis_token_get_text(t);
-	if ((t->type != SUBTILIS_TOKEN_OPERATOR) || strcmp(tbuf, ")")) {
-		subtilis_error_set_right_bkt_expected(
-		    err, tbuf, p->l->stream->name, p->l->line);
-		return NULL;
-	}
-
-	return e;
-}
-
-static subtilis_exp_t *prv_unary_minus_exp(subtilis_parser_t *p,
-					   subtilis_token_t *t,
-					   subtilis_error_t *err)
-{
-	subtilis_exp_t *e;
-
-	subtilis_lexer_get(p->l, t, err);
-	if (err->type != SUBTILIS_ERROR_OK)
-		return NULL;
-
-	e = prv_priority1(p, t, err);
-	if (err->type != SUBTILIS_ERROR_OK)
-		return NULL;
-	return subtilis_type_if_unary_minus(p, e, err);
-}
-
-static subtilis_exp_t *prv_not_exp(subtilis_parser_t *p, subtilis_token_t *t,
-				   subtilis_error_t *err)
-{
-	subtilis_exp_t *e;
-
-	subtilis_lexer_get(p->l, t, err);
-	if (err->type != SUBTILIS_ERROR_OK)
-		return NULL;
-
-	e = prv_priority1(p, t, err);
-	if (err->type != SUBTILIS_ERROR_OK)
-		return NULL;
-	return subtilis_type_if_not(p, e, err);
-}
-
-static subtilis_exp_t *prv_gettime(subtilis_parser_t *p, subtilis_token_t *t,
-				   subtilis_error_t *err)
-{
-	size_t reg;
-	subtilis_exp_t *e;
-
-	reg = subtilis_ir_section_add_instr1(p->current,
-					     SUBTILIS_OP_INSTR_GETTIME, err);
-	if (err->type != SUBTILIS_ERROR_OK)
-		return NULL;
-
-	subtilis_exp_handle_errors(p, err);
-	if (err->type != SUBTILIS_ERROR_OK)
-		return NULL;
-
-	e = subtilis_exp_new_int32_var(reg, err);
-	if (err->type != SUBTILIS_ERROR_OK)
-		return NULL;
-
-	subtilis_lexer_get(p->l, t, err);
-	if (err->type != SUBTILIS_ERROR_OK)
-		goto on_error;
-
-	return e;
-
-on_error:
-	subtilis_exp_delete(e);
-	return NULL;
-}
-
-static subtilis_exp_t *prv_parse_bracketed_exp(subtilis_parser_t *p,
-					       subtilis_token_t *t,
-					       subtilis_error_t *err)
-{
-	const char *tbuf;
-	subtilis_exp_t *e = NULL;
-
-	subtilis_lexer_get(p->l, t, err);
-	if (err->type != SUBTILIS_ERROR_OK)
-		return NULL;
-
-	tbuf = subtilis_token_get_text(t);
-	if (strcmp(tbuf, "(")) {
-		subtilis_error_set_exp_expected(err, "( ", p->l->stream->name,
-						p->l->line);
-		return NULL;
-	}
-
-	e = prv_bracketed_exp(p, t, err);
-	if (err->type != SUBTILIS_ERROR_OK)
-		return NULL;
-
-	subtilis_lexer_get(p->l, t, err);
-	if (err->type != SUBTILIS_ERROR_OK)
-		goto cleanup;
-
-	return e;
-
-cleanup:
-
-	subtilis_exp_delete(e);
-	return NULL;
-}
-
-static subtilis_exp_t *prv_real_bracketed_exp(subtilis_parser_t *p,
-					      subtilis_token_t *t,
-					      subtilis_error_t *err)
-{
-	subtilis_exp_t *e;
-
-	e = prv_parse_bracketed_exp(p, t, err);
-	if (err->type != SUBTILIS_ERROR_OK)
-		return NULL;
-
-	return subtilis_type_if_to_float64(p, e, err);
-}
-
-static subtilis_exp_t *prv_real_unary_fn(subtilis_parser_t *p,
-					 subtilis_token_t *t,
-					 subtilis_op_instr_type_t itype,
-					 double (*const_fn)(double),
-					 subtilis_error_t *err)
-{
-	size_t reg;
-	subtilis_exp_t *e = NULL;
-	subtilis_exp_t *e2 = NULL;
-
-	e = prv_real_bracketed_exp(p, t, err);
-	if (err->type != SUBTILIS_ERROR_OK)
-		return NULL;
-
-	if (e->type.type == SUBTILIS_TYPE_CONST_REAL) {
-		e2 = subtilis_exp_new_real(const_fn(e->exp.ir_op.real), err);
-	} else {
-		reg = subtilis_ir_section_add_instr2(p->current, itype,
-						     e->exp.ir_op, err);
-		if (err->type != SUBTILIS_ERROR_OK)
-			goto cleanup;
-
-		e2 = subtilis_exp_new_real_var(reg, err);
-	}
-	if (err->type != SUBTILIS_ERROR_OK)
-		goto cleanup;
-
-	subtilis_exp_delete(e);
-
-	return e2;
-
-cleanup:
-
-	subtilis_exp_delete(e);
-	return NULL;
-}
-
-static subtilis_exp_t *prv_integer_bracketed_exp(subtilis_parser_t *p,
-						 subtilis_token_t *t,
-						 subtilis_error_t *err)
-{
-	subtilis_exp_t *e;
-
-	e = prv_parse_bracketed_exp(p, t, err);
-	if (err->type != SUBTILIS_ERROR_OK)
-		return NULL;
-
-	return subtilis_type_if_to_int(p, e, err);
-}
-
-static subtilis_exp_t *prv_rad(subtilis_parser_t *p, subtilis_token_t *t,
-			       subtilis_error_t *err)
-{
-	subtilis_exp_t *e = NULL;
-	subtilis_exp_t *e2 = NULL;
-	subtilis_exp_t *e3 = NULL;
-	const double radian = ((22 / 7.0) / 180);
-
-	e = prv_real_bracketed_exp(p, t, err);
-	if (err->type != SUBTILIS_ERROR_OK)
-		return NULL;
-
-	if (e->type.type == SUBTILIS_TYPE_CONST_REAL) {
-		e2 = subtilis_exp_new_real(e->exp.ir_op.real * radian, err);
-	} else {
-		e3 = subtilis_exp_new_real(radian, err);
-		if (err->type != SUBTILIS_ERROR_OK)
-			goto cleanup;
-		e2 = subtilis_type_if_mul(p, e, e3, err);
-		e = NULL;
-	}
-	if (err->type != SUBTILIS_ERROR_OK)
-		goto cleanup;
-
-	subtilis_exp_delete(e);
-
-	return e2;
-
-cleanup:
-
-	subtilis_exp_delete(e);
-	return NULL;
-}
-
-static subtilis_exp_t *prv_pi(subtilis_parser_t *p, subtilis_token_t *t,
-			      subtilis_error_t *err)
-{
-	subtilis_exp_t *e;
-
-	e = subtilis_exp_new_real(22 / 7.0, err);
-	if (err->type != SUBTILIS_ERROR_OK)
-		return NULL;
-
-	subtilis_lexer_get(p->l, t, err);
-	if (err->type != SUBTILIS_ERROR_OK) {
-		subtilis_exp_delete(e);
-		return NULL;
-	}
-
-	return e;
-}
-
-static subtilis_exp_t *prv_int(subtilis_parser_t *p, subtilis_token_t *t,
-			       subtilis_error_t *err)
-{
-	size_t reg;
-	subtilis_exp_t *e = NULL;
-	subtilis_exp_t *e2 = NULL;
-
-	e = prv_real_bracketed_exp(p, t, err);
-	if (err->type != SUBTILIS_ERROR_OK)
-		return NULL;
-
-	if (e->type.type == SUBTILIS_TYPE_CONST_REAL) {
-		e2 = subtilis_exp_new_int32(floor(e->exp.ir_op.real), err);
-	} else {
-		reg = subtilis_ir_section_add_instr2(
-		    p->current, SUBTILIS_OP_INSTR_MOV_FPRD_I32, e->exp.ir_op,
-		    err);
-		if (err->type != SUBTILIS_ERROR_OK)
-			goto cleanup;
-
-		e2 = subtilis_exp_new_int32_var(reg, err);
-	}
-	if (err->type != SUBTILIS_ERROR_OK)
-		goto cleanup;
-
-	subtilis_exp_delete(e);
-
-	return e2;
-
-cleanup:
-
-	subtilis_exp_delete(e);
-	return NULL;
-}
-
-static subtilis_exp_t *prv_sin(subtilis_parser_t *p, subtilis_token_t *t,
-			       subtilis_error_t *err)
-{
-	return prv_real_unary_fn(p, t, SUBTILIS_OP_INSTR_SIN, sin, err);
-}
-
-static subtilis_exp_t *prv_cos(subtilis_parser_t *p, subtilis_token_t *t,
-			       subtilis_error_t *err)
-{
-	return prv_real_unary_fn(p, t, SUBTILIS_OP_INSTR_COS, cos, err);
-}
-
-static subtilis_exp_t *prv_tan(subtilis_parser_t *p, subtilis_token_t *t,
-			       subtilis_error_t *err)
-{
-	return prv_real_unary_fn(p, t, SUBTILIS_OP_INSTR_TAN, tan, err);
-}
-
-static subtilis_exp_t *prv_asn(subtilis_parser_t *p, subtilis_token_t *t,
-			       subtilis_error_t *err)
-{
-	return prv_real_unary_fn(p, t, SUBTILIS_OP_INSTR_ASN, asin, err);
-}
-
-static subtilis_exp_t *prv_acs(subtilis_parser_t *p, subtilis_token_t *t,
-			       subtilis_error_t *err)
-{
-	return prv_real_unary_fn(p, t, SUBTILIS_OP_INSTR_ACS, acos, err);
-}
-
-static subtilis_exp_t *prv_atn(subtilis_parser_t *p, subtilis_token_t *t,
-			       subtilis_error_t *err)
-{
-	return prv_real_unary_fn(p, t, SUBTILIS_OP_INSTR_ATN, atan, err);
-}
-
-static subtilis_exp_t *prv_log(subtilis_parser_t *p, subtilis_token_t *t,
-			       subtilis_op_instr_type_t itype,
-			       double (*log_fn)(double), subtilis_error_t *err)
-{
-	subtilis_exp_t *e;
-
-	e = prv_real_unary_fn(p, t, itype, log_fn, err);
-	if (err->type != SUBTILIS_ERROR_OK)
-		return NULL;
-	subtilis_exp_handle_errors(p, err);
-	if (err->type != SUBTILIS_ERROR_OK) {
-		subtilis_exp_delete(e);
-		return NULL;
-	}
-	return e;
-}
-
-static subtilis_exp_t *prv_rnd_var(subtilis_parser_t *p, subtilis_exp_t *e,
-				   subtilis_error_t *err)
-{
-	size_t reg;
-	subtilis_ir_section_t *fn;
-	size_t zero;
-	size_t one;
-	size_t cond;
-	subtilis_ir_operand_t op1;
-	subtilis_ir_operand_t op2;
-	subtilis_ir_operand_t else_label;
-	subtilis_ir_operand_t zero_one_label;
-	subtilis_ir_operand_t end_label;
-	subtilis_ir_operand_t result;
-
-	zero_one_label.label = subtilis_ir_section_new_label(p->current);
-	else_label.label = subtilis_ir_section_new_label(p->current);
-	end_label.label = subtilis_ir_section_new_label(p->current);
-
-	result.reg = p->current->freg_counter++;
-	reg = e->exp.ir_op.reg;
-	subtilis_exp_delete(e);
-	e = NULL;
-
-	fn = subtilis_builtins_ir_add_1_arg_int(p, "_rnd_int",
-						&subtilis_type_integer, err);
-	if (err->type != SUBTILIS_ERROR_OK) {
-		if (err->type != SUBTILIS_ERROR_ALREADY_DEFINED)
-			goto cleanup;
-		subtilis_error_init(err);
-	} else {
-		subtilis_builtins_ir_rnd_int(p, fn, err);
-		if (err->type != SUBTILIS_ERROR_OK)
-			goto cleanup;
-	}
-
-	fn = subtilis_builtins_ir_add_1_arg_int(p, "_rnd_real",
-						&subtilis_type_real, err);
-	if (err->type != SUBTILIS_ERROR_OK) {
-		if (err->type != SUBTILIS_ERROR_ALREADY_DEFINED)
-			goto cleanup;
-		subtilis_error_init(err);
-	} else {
-		subtilis_builtins_ir_rnd_real(p, fn, err);
-		if (err->type != SUBTILIS_ERROR_OK)
-			goto cleanup;
-	}
-
-	op1.reg = reg;
-	op2.integer = 0;
-	zero = subtilis_ir_section_add_instr(
-	    p->current, SUBTILIS_OP_INSTR_EQI_I32, op1, op2, err);
-	if (err->type != SUBTILIS_ERROR_OK)
-		goto cleanup;
-
-	op2.integer = 1;
-	one = subtilis_ir_section_add_instr(
-	    p->current, SUBTILIS_OP_INSTR_EQI_I32, op1, op2, err);
-	if (err->type != SUBTILIS_ERROR_OK)
-		goto cleanup;
-
-	op1.reg = zero;
-	op2.reg = one;
-	cond = subtilis_ir_section_add_instr(
-	    p->current, SUBTILIS_OP_INSTR_OR_I32, op1, op2, err);
-	if (err->type != SUBTILIS_ERROR_OK)
-		goto cleanup;
-
-	op1.reg = cond;
-	subtilis_ir_section_add_instr_reg(p->current, SUBTILIS_OP_INSTR_JMPC,
-					  op1, zero_one_label, else_label, err);
-	if (err->type != SUBTILIS_ERROR_OK)
-		goto cleanup;
-
-	subtilis_ir_section_add_label(p->current, zero_one_label.reg, err);
-	if (err->type != SUBTILIS_ERROR_OK)
-		goto cleanup;
-
-	e = prv_call_1_arg_fn(p, "_rnd_real", reg, SUBTILIS_IR_REG_TYPE_INTEGER,
-			      &subtilis_type_real, err);
-	if (err->type != SUBTILIS_ERROR_OK)
-		goto cleanup;
-
-	subtilis_ir_section_add_instr_no_reg2(
-	    p->current, SUBTILIS_OP_INSTR_MOVFP, result, e->exp.ir_op, err);
-	if (err->type != SUBTILIS_ERROR_OK)
-		goto cleanup;
-
-	subtilis_exp_delete(e);
-	e = NULL;
-
-	subtilis_ir_section_add_instr_no_reg(p->current, SUBTILIS_OP_INSTR_JMP,
-					     end_label, err);
-	if (err->type != SUBTILIS_ERROR_OK)
-		goto cleanup;
-
-	subtilis_ir_section_add_label(p->current, else_label.reg, err);
-	if (err->type != SUBTILIS_ERROR_OK)
-		goto cleanup;
-
-	e = prv_call_1_arg_fn(p, "_rnd_int", reg, SUBTILIS_IR_REG_TYPE_INTEGER,
-			      &subtilis_type_integer, err);
-	if (err->type != SUBTILIS_ERROR_OK)
-		goto cleanup;
-
-	subtilis_ir_section_add_instr_no_reg2(p->current,
-					      SUBTILIS_OP_INSTR_MOV_I32_FP,
-					      result, e->exp.ir_op, err);
-	if (err->type != SUBTILIS_ERROR_OK)
-		goto cleanup;
-
-	subtilis_exp_delete(e);
-	e = NULL;
-
-	subtilis_ir_section_add_label(p->current, end_label.reg, err);
-	if (err->type != SUBTILIS_ERROR_OK)
-		goto cleanup;
-
-	return subtilis_exp_new_real_var(result.reg, err);
-
-cleanup:
-
-	subtilis_exp_delete(e);
-
-	return NULL;
-}
-
-static subtilis_exp_t *prv_rnd_const(subtilis_parser_t *p, subtilis_exp_t *e,
-				     subtilis_error_t *err)
-{
-	int32_t val = e->exp.ir_op.integer;
-
-	subtilis_exp_delete(e);
-	if (val == 0)
-		return subtilis_builtins_ir_rnd_0(p, err);
-	else if (val == 1)
-		return subtilis_builtins_ir_rnd_1(p, err);
-	else if (val < 0)
-		return subtilis_builtins_ir_rnd_neg(p, val, err);
-	else
-		return subtilis_builtins_ir_rnd_pos(p, val, err);
-}
-
-static subtilis_exp_t *prv_rnd(subtilis_parser_t *p, subtilis_token_t *t,
-			       subtilis_error_t *err)
-{
-	const char *tbuf;
-	subtilis_exp_t *e = NULL;
-	subtilis_exp_t *e_dup = NULL;
-
-	subtilis_lexer_get(p->l, t, err);
-	if (err->type != SUBTILIS_ERROR_OK)
-		goto cleanup;
-
-	if (t->type == SUBTILIS_TOKEN_OPERATOR) {
-		tbuf = subtilis_token_get_text(t);
-		if (!strcmp(tbuf, "(")) {
-			e = prv_bracketed_exp(p, t, err);
-			if (err->type != SUBTILIS_ERROR_OK)
-				goto cleanup;
-
-			subtilis_lexer_get(p->l, t, err);
-			if (err->type != SUBTILIS_ERROR_OK)
-				goto cleanup;
-
-			e = subtilis_type_if_to_int(p, e, err);
-			if (err->type != SUBTILIS_ERROR_OK)
-				goto cleanup;
-
-			if (e->type.type == SUBTILIS_TYPE_CONST_INTEGER)
-				return prv_rnd_const(p, e, err);
-			return prv_rnd_var(p, e, err);
-		}
-	}
-
-	e = subtilis_builtins_ir_basic_rnd(p, err);
-	if (err->type != SUBTILIS_ERROR_OK)
-		goto cleanup;
-
-	e_dup = subtilis_type_if_dup(e, err);
-	if (err->type != SUBTILIS_ERROR_OK)
-		goto cleanup;
-
-	subtilis_var_assign_hidden(p, subtilis_rnd_hidden_var,
-				   &subtilis_type_integer, e_dup, err);
-	if (err->type != SUBTILIS_ERROR_OK)
-		goto cleanup;
-
-	return e;
-
-cleanup:
-
-	subtilis_exp_delete(e);
-	subtilis_exp_delete(e_dup);
-
-	return NULL;
-}
-
-static subtilis_exp_t *prv_sqr(subtilis_parser_t *p, subtilis_token_t *t,
-			       subtilis_error_t *err)
-{
-	return prv_real_unary_fn(p, t, SUBTILIS_OP_INSTR_SQR, sqrt, err);
-}
-
-static subtilis_exp_t *prv_get(subtilis_parser_t *p, subtilis_token_t *t,
-			       subtilis_error_t *err)
-{
-	size_t reg;
-	subtilis_exp_t *e;
-
-	reg = subtilis_ir_section_add_instr1(p->current, SUBTILIS_OP_INSTR_GET,
-					     err);
-	if (err->type != SUBTILIS_ERROR_OK)
-		return NULL;
-
-	subtilis_exp_handle_errors(p, err);
-	if (err->type != SUBTILIS_ERROR_OK)
-		return NULL;
-
-	e = subtilis_exp_new_int32_var(reg, err);
-
-	subtilis_lexer_get(p->l, t, err);
-	if (err->type != SUBTILIS_ERROR_OK) {
-		subtilis_exp_delete(e);
-		return NULL;
-	}
-
-	return e;
-}
-
-static subtilis_exp_t *prv_inkey_const(subtilis_parser_t *p,
-				       subtilis_token_t *t, subtilis_exp_t *e,
-				       subtilis_error_t *err)
-{
-	size_t reg;
-	subtilis_op_instr_type_t itype;
-	subtilis_ir_operand_t op1;
-	int32_t inkey_value = e->exp.ir_op.integer;
-
-	if (inkey_value == -256) {
-		reg = subtilis_ir_section_add_instr1(
-		    p->current, SUBTILIS_OP_INSTR_OS_BYTE_ID, err);
-		if (err->type != SUBTILIS_ERROR_OK)
-			goto cleanup;
-		subtilis_exp_handle_errors(p, err);
-		if (err->type != SUBTILIS_ERROR_OK)
-			goto cleanup;
-	} else {
-		if (inkey_value >= 0)
-			itype = SUBTILIS_OP_INSTR_GET_TO;
-		else
-			itype = SUBTILIS_OP_INSTR_INKEY;
-
-		e = subtilis_type_if_exp_to_var(p, e, err);
-		if (err->type != SUBTILIS_ERROR_OK)
-			goto cleanup;
-
-		op1.reg = e->exp.ir_op.reg;
-		;
-		reg =
-		    subtilis_ir_section_add_instr2(p->current, itype, op1, err);
-		if (err->type != SUBTILIS_ERROR_OK)
-			goto cleanup;
-		subtilis_exp_handle_errors(p, err);
-		if (err->type != SUBTILIS_ERROR_OK)
-			goto cleanup;
-	}
-
-	subtilis_exp_delete(e);
-
-	return subtilis_exp_new_int32_var(reg, err);
-
-cleanup:
-
-	subtilis_exp_delete(e);
-	return NULL;
-}
-
-static subtilis_exp_t *prv_call_1_arg_fn(subtilis_parser_t *p, const char *name,
-					 size_t reg,
-					 subtilis_ir_reg_type_t ptype,
-					 const subtilis_type_t *rtype,
-					 subtilis_error_t *err)
-{
-	subtilis_ir_arg_t *args = NULL;
-	char *name_dup = NULL;
-
-	args = malloc(sizeof(*args) * 1);
-	if (!args) {
-		subtilis_error_set_oom(err);
-		return NULL;
-	}
-
-	name_dup = malloc(strlen(name) + 1);
-	if (!name_dup) {
-		subtilis_error_set_oom(err);
-		goto cleanup;
-	}
-	strcpy(name_dup, name);
-
-	args[0].type = ptype;
-	args[0].reg = reg;
-
-	return subtilis_exp_add_call(p, name_dup, SUBTILIS_BUILTINS_MAX, NULL,
-				     args, rtype, 1, err);
-
-cleanup:
-
-	free(args);
-
-	return NULL;
-}
-
-static subtilis_exp_t *prv_inkey(subtilis_parser_t *p, subtilis_token_t *t,
-				 subtilis_error_t *err)
-{
-	subtilis_exp_t *e;
-	size_t reg;
-	subtilis_ir_section_t *current;
-
-	e = prv_integer_bracketed_exp(p, t, err);
-	if (err->type != SUBTILIS_ERROR_OK)
-		return NULL;
-
-	if (e->type.type == SUBTILIS_TYPE_CONST_INTEGER)
-		return prv_inkey_const(p, t, e, err);
-
-	reg = e->exp.ir_op.reg;
-	subtilis_exp_delete(e);
-
-	current = subtilis_builtins_ir_add_1_arg_int(
-	    p, "_inkey", &subtilis_type_integer, err);
-	if (err->type != SUBTILIS_ERROR_OK) {
-		if (err->type != SUBTILIS_ERROR_ALREADY_DEFINED)
-			return NULL;
-		subtilis_error_init(err);
-	} else {
-		subtilis_builtins_ir_inkey(current, err);
-		if (err->type != SUBTILIS_ERROR_OK)
-			return NULL;
-	}
-
-	return prv_call_1_arg_fn(p, "_inkey", reg, SUBTILIS_IR_REG_TYPE_INTEGER,
-				 &subtilis_type_integer, err);
-}
-
-static subtilis_exp_t *prv_abs(subtilis_parser_t *p, subtilis_token_t *t,
-			       subtilis_error_t *err)
-{
-	subtilis_exp_t *e;
-	size_t reg;
-	subtilis_ir_operand_t op1;
-	subtilis_ir_operand_t op2;
-
-	e = prv_parse_bracketed_exp(p, t, err);
-	if (err->type != SUBTILIS_ERROR_OK)
-		return NULL;
-
-	switch (e->type.type) {
-	case SUBTILIS_TYPE_CONST_INTEGER:
-		e->exp.ir_op.integer = abs(e->exp.ir_op.integer);
-		break;
-	case SUBTILIS_TYPE_CONST_REAL:
-		e->exp.ir_op.real = fabs(e->exp.ir_op.real);
-		break;
-	case SUBTILIS_TYPE_INTEGER:
-		op2.integer = 31;
-		reg = subtilis_ir_section_add_instr(p->current,
-						    SUBTILIS_OP_INSTR_ASRI_I32,
-						    e->exp.ir_op, op2, err);
-		if (err->type != SUBTILIS_ERROR_OK)
-			goto cleanup;
-		op2.reg = reg;
-		reg = subtilis_ir_section_add_instr(p->current,
-						    SUBTILIS_OP_INSTR_ADD_I32,
-						    e->exp.ir_op, op2, err);
-		if (err->type != SUBTILIS_ERROR_OK)
-			goto cleanup;
-		op1.reg = reg;
-		reg = subtilis_ir_section_add_instr(
-		    p->current, SUBTILIS_OP_INSTR_EOR_I32, op1, op2, err);
-		if (err->type != SUBTILIS_ERROR_OK)
-			goto cleanup;
-		e->exp.ir_op.reg = reg;
-		break;
-	case SUBTILIS_TYPE_REAL:
-		reg = subtilis_ir_section_add_instr2(
-		    p->current, SUBTILIS_OP_INSTR_ABSR, e->exp.ir_op, err);
-		if (err->type != SUBTILIS_ERROR_OK)
-			goto cleanup;
-
-		subtilis_exp_delete(e);
-		e = subtilis_exp_new_real_var(reg, err);
-		break;
-	default:
-		subtilis_error_set_numeric_expected(err, "", p->l->stream->name,
-						    p->l->line);
-		goto cleanup;
-	}
-
-	return e;
-
-cleanup:
-
-	subtilis_exp_delete(e);
-	return NULL;
-}
-
-static subtilis_exp_t *prv_get_point(subtilis_parser_t *p, subtilis_token_t *t,
-				     subtilis_error_t *err)
-{
-	return prv_bracketed_2_int_args(p, t, SUBTILIS_OP_INSTR_POINT, err);
-}
-
-static subtilis_exp_t *prv_get_tint(subtilis_parser_t *p, subtilis_token_t *t,
-				    subtilis_error_t *err)
-{
-	return prv_bracketed_2_int_args(p, t, SUBTILIS_OP_INSTR_TINT, err);
-}
-
-static subtilis_exp_t *prv_get_err(subtilis_parser_t *p, subtilis_token_t *t,
-				   subtilis_error_t *err)
-{
-	subtilis_exp_t *e;
-
-	e = subtilis_var_lookup_var(p, subtilis_err_hidden_var, err);
-	if (err->type != SUBTILIS_ERROR_OK)
-		return NULL;
-
-	subtilis_lexer_get(p->l, t, err);
-	if (err->type != SUBTILIS_ERROR_OK) {
-		subtilis_exp_delete(e);
-		return NULL;
-	}
-
-	return e;
-}
-
-static subtilis_exp_t *prv_read_array(subtilis_parser_t *p, subtilis_token_t *t,
-				      const char *var_name,
-				      subtilis_error_t *err)
-{
-	size_t mem_reg;
-	const subtilis_symbol_t *s;
-	subtilis_exp_t *indices[SUBTILIS_MAX_DIMENSIONS];
-	size_t i;
-	subtilis_exp_t *e = NULL;
-	size_t dims = 0;
-
-	s = subtilis_symbol_table_lookup(p->local_st, var_name);
-	if (s) {
-		mem_reg = SUBTILIS_IR_REG_LOCAL;
-	} else {
-		s = subtilis_symbol_table_lookup(p->st, var_name);
-		if (!s) {
-			subtilis_error_set_unknown_variable(
-			    err, var_name, p->l->stream->name, p->l->line);
-			return NULL;
-		}
-		mem_reg = SUBTILIS_IR_REG_GLOBAL;
-	}
-
-	dims = prv_var_bracketed_int_args_have_b(p, t, indices,
-						 SUBTILIS_MAX_DIMENSIONS, err);
-	if (err->type != SUBTILIS_ERROR_OK)
-		return NULL;
-
-	subtilis_lexer_get(p->l, t, err);
-	if (err->type != SUBTILIS_ERROR_OK)
-		goto cleanup;
-
-	e = subtilis_type_if_indexed_read(p, var_name, &s->t, mem_reg, s->loc,
-					  indices, dims, err);
-
-cleanup:
-
-	for (i = 0; i < dims; i++)
-		subtilis_exp_delete(indices[i]);
-
-	return e;
-}
-
-static subtilis_exp_t *prv_lookup_var(subtilis_parser_t *p, subtilis_token_t *t,
-				      const char *tbuf, subtilis_error_t *err)
-{
-	subtilis_exp_t *e = NULL;
-	char *var_name = NULL;
-
-	tbuf = subtilis_token_get_text(t);
-	var_name = malloc(strlen(tbuf) + 1);
-	if (!var_name) {
-		subtilis_error_set_oom(err);
-		return NULL;
-	}
-	strcpy(var_name, tbuf);
-
-	subtilis_lexer_get(p->l, t, err);
-	if (err->type != SUBTILIS_ERROR_OK)
-		goto cleanup;
-
-	tbuf = subtilis_token_get_text(t);
-	if ((t->type != SUBTILIS_TOKEN_OPERATOR) || (strcmp(tbuf, "(")))
-		e = subtilis_var_lookup_var(p, var_name, err);
-	else
-		e = prv_read_array(p, t, var_name, err);
-
-cleanup:
-
-	free(var_name);
-
-	return e;
-}
-
-static subtilis_exp_t *prv_priority1(subtilis_parser_t *p, subtilis_token_t *t,
-				     subtilis_error_t *err)
-{
-	const char *tbuf;
-	subtilis_exp_t *e = NULL;
-
-	tbuf = subtilis_token_get_text(t);
-	switch (t->type) {
-	case SUBTILIS_TOKEN_INTEGER:
-		e = subtilis_exp_new_int32(t->tok.integer, err);
-		if (!e)
-			goto cleanup;
-		break;
-	case SUBTILIS_TOKEN_REAL:
-		e = subtilis_exp_new_real(t->tok.real, err);
-		if (!e)
-			goto cleanup;
-		break;
-	case SUBTILIS_TOKEN_OPERATOR:
-		if (!strcmp(tbuf, "(")) {
-			e = prv_bracketed_exp(p, t, err);
-			if (err->type != SUBTILIS_ERROR_OK)
-				goto cleanup;
-		} else {
-			if (!strcmp(tbuf, "-")) {
-				e = prv_unary_minus_exp(p, t, err);
-				if (err->type != SUBTILIS_ERROR_OK)
-					goto cleanup;
-
-				/* we don't want to read another token here.It's
-				 * already been read by the recursive call to
-				 * prv_priority1.
-				 */
-
-				return e;
-			}
-			subtilis_error_set_exp_expected(
-			    err, "( or - ", p->l->stream->name, p->l->line);
-			goto cleanup;
-		}
-		break;
-	case SUBTILIS_TOKEN_IDENTIFIER:
-		e = prv_lookup_var(p, t, tbuf, err);
-		if (err->type != SUBTILIS_ERROR_OK)
-			goto cleanup;
-		return e;
-	case SUBTILIS_TOKEN_KEYWORD:
-		switch (t->tok.keyword.type) {
-		case SUBTILIS_KEYWORD_TRUE:
-			e = subtilis_exp_new_int32(-1, err);
-			if (!e)
-				goto cleanup;
-			break;
-		case SUBTILIS_KEYWORD_FALSE:
-			e = subtilis_exp_new_int32(0, err);
-			if (!e)
-				goto cleanup;
-			break;
-		case SUBTILIS_KEYWORD_NOT:
-			e = prv_not_exp(p, t, err);
-			if (err->type != SUBTILIS_ERROR_OK)
-				goto cleanup;
-			return e;
-		case SUBTILIS_KEYWORD_FN:
-			return prv_call(p, t, err);
-		case SUBTILIS_KEYWORD_INT:
-			return prv_int(p, t, err);
-		case SUBTILIS_KEYWORD_TIME:
-			return prv_gettime(p, t, err);
-		case SUBTILIS_KEYWORD_SIN:
-			return prv_sin(p, t, err);
-		case SUBTILIS_KEYWORD_COS:
-			return prv_cos(p, t, err);
-		case SUBTILIS_KEYWORD_TAN:
-			return prv_tan(p, t, err);
-		case SUBTILIS_KEYWORD_ACS:
-			return prv_acs(p, t, err);
-		case SUBTILIS_KEYWORD_ASN:
-			return prv_asn(p, t, err);
-		case SUBTILIS_KEYWORD_ATN:
-			return prv_atn(p, t, err);
-		case SUBTILIS_KEYWORD_RND:
-			return prv_rnd(p, t, err);
-		case SUBTILIS_KEYWORD_SQR:
-			return prv_sqr(p, t, err);
-		case SUBTILIS_KEYWORD_LOG:
-			return prv_log(p, t, SUBTILIS_OP_INSTR_LOG, log10, err);
-		case SUBTILIS_KEYWORD_LN:
-			return prv_log(p, t, SUBTILIS_OP_INSTR_LN, log, err);
-		case SUBTILIS_KEYWORD_RAD:
-			return prv_rad(p, t, err);
-		case SUBTILIS_KEYWORD_ABS:
-			return prv_abs(p, t, err);
-		case SUBTILIS_KEYWORD_PI:
-			return prv_pi(p, t, err);
-		case SUBTILIS_KEYWORD_GET:
-			return prv_get(p, t, err);
-		case SUBTILIS_KEYWORD_INKEY:
-			return prv_inkey(p, t, err);
-		case SUBTILIS_KEYWORD_POINT:
-			return prv_get_point(p, t, err);
-		case SUBTILIS_KEYWORD_TINT:
-			return prv_get_tint(p, t, err);
-		case SUBTILIS_KEYWORD_ERR:
-			return prv_get_err(p, t, err);
-		default:
-			subtilis_error_set_exp_expected(
-			    err, "Unexpected keyword in expression",
-			    p->l->stream->name, p->l->line);
-			goto cleanup;
-		}
-		break;
-	default:
-		subtilis_error_set_exp_expected(err, tbuf, p->l->stream->name,
-						p->l->line);
-		goto cleanup;
-	}
-
-	subtilis_lexer_get(p->l, t, err);
-	if (err->type != SUBTILIS_ERROR_OK)
-		goto cleanup;
-
-	return e;
-
-cleanup:
-
-	subtilis_exp_delete(e);
-	return NULL;
-}
-
-static subtilis_exp_t *prv_priority2(subtilis_parser_t *p, subtilis_token_t *t,
-				     subtilis_error_t *err)
-{
-	return prv_priority1(p, t, err);
-}
-
-static subtilis_exp_t *prv_priority3(subtilis_parser_t *p, subtilis_token_t *t,
-				     subtilis_error_t *err)
-{
-	const char *tbuf;
-	subtilis_exp_t *e1 = NULL;
-	subtilis_exp_t *e2 = NULL;
-	subtilis_exp_fn_t exp_fn;
-
-	e1 = prv_priority2(p, t, err);
-	if (err->type != SUBTILIS_ERROR_OK)
-		goto cleanup;
-	while ((t->type == SUBTILIS_TOKEN_OPERATOR) ||
-	       (t->type == SUBTILIS_TOKEN_KEYWORD)) {
-		tbuf = subtilis_token_get_text(t);
-		if (t->type == SUBTILIS_TOKEN_KEYWORD) {
-			if (!strcmp(tbuf, "DIV"))
-				exp_fn = subtilis_type_if_div;
-			else if (!strcmp(tbuf, "MOD"))
-				exp_fn = subtilis_type_if_mod;
-			else
-				break;
-		} else if (!strcmp(tbuf, "*")) {
-			exp_fn = subtilis_type_if_mul;
-		} else if (!strcmp(tbuf, "/")) {
-			exp_fn = subtilis_type_if_divr;
-		} else {
-			break;
-		}
-		subtilis_lexer_get(p->l, t, err);
-		if (err->type != SUBTILIS_ERROR_OK)
-			goto cleanup;
-		e2 = prv_priority2(p, t, err);
-		if (err->type != SUBTILIS_ERROR_OK)
-			goto cleanup;
-
-		e1 = exp_fn(p, e1, e2, err);
-		e2 = NULL;
-		if (err->type != SUBTILIS_ERROR_OK)
-			goto cleanup;
-	}
-
-	return e1;
-
-cleanup:
-
-	subtilis_exp_delete(e2);
-	subtilis_exp_delete(e1);
-	return NULL;
-}
-
-static subtilis_exp_t *prv_priority4(subtilis_parser_t *p, subtilis_token_t *t,
-				     subtilis_error_t *err)
-{
-	const char *tbuf;
-	subtilis_exp_t *e1 = NULL;
-	subtilis_exp_t *e2 = NULL;
-	subtilis_exp_fn_t exp_fn;
-
-	e1 = prv_priority3(p, t, err);
-	if (err->type != SUBTILIS_ERROR_OK)
-		goto cleanup;
-	tbuf = subtilis_token_get_text(t);
-	while (t->type == SUBTILIS_TOKEN_OPERATOR) {
-		tbuf = subtilis_token_get_text(t);
-		if (!strcmp(tbuf, "+"))
-			exp_fn = subtilis_exp_add;
-		else if (!strcmp(tbuf, "-"))
-			exp_fn = subtilis_type_if_sub;
-		else
-			break;
-		subtilis_lexer_get(p->l, t, err);
-		if (err->type != SUBTILIS_ERROR_OK)
-			goto cleanup;
-		e2 = prv_priority3(p, t, err);
-		if (err->type != SUBTILIS_ERROR_OK)
-			goto cleanup;
-
-		e1 = exp_fn(p, e1, e2, err);
-		e2 = NULL;
-		if (err->type != SUBTILIS_ERROR_OK)
-			goto cleanup;
-	}
-
-	return e1;
-
-cleanup:
-
-	subtilis_exp_delete(e2);
-	subtilis_exp_delete(e1);
-	return NULL;
-}
-
-static subtilis_exp_t *prv_priority5(subtilis_parser_t *p, subtilis_token_t *t,
-				     subtilis_error_t *err)
-{
-	const char *tbuf;
-	subtilis_exp_t *e1 = NULL;
-	subtilis_exp_t *e2 = NULL;
-	subtilis_exp_fn_t exp_fn;
-
-	e1 = prv_priority4(p, t, err);
-	if (err->type != SUBTILIS_ERROR_OK)
-		goto cleanup;
-
-	while (t->type == SUBTILIS_TOKEN_OPERATOR) {
-		tbuf = subtilis_token_get_text(t);
-		if (!strcmp(tbuf, "="))
-			exp_fn = subtilis_type_if_eq;
-		else if (!strcmp(tbuf, "<>"))
-			exp_fn = subtilis_type_if_neq;
-		else if (!strcmp(tbuf, ">"))
-			exp_fn = subtilis_exp_gt;
-		else if (!strcmp(tbuf, "<="))
-			exp_fn = subtilis_exp_lte;
-		else if (!strcmp(tbuf, "<"))
-			exp_fn = subtilis_exp_lt;
-		else if (!strcmp(tbuf, ">="))
-			exp_fn = subtilis_exp_gte;
-		else if (!strcmp(tbuf, "<<"))
-			exp_fn = subtilis_type_if_lsl;
-		else if (!strcmp(tbuf, ">>"))
-			exp_fn = subtilis_type_if_lsr;
-		else if (!strcmp(tbuf, ">>>"))
-			exp_fn = subtilis_type_if_asr;
-		else
-			break;
-		subtilis_lexer_get(p->l, t, err);
-		if (err->type != SUBTILIS_ERROR_OK)
-			goto cleanup;
-		e2 = prv_priority4(p, t, err);
-		if (err->type != SUBTILIS_ERROR_OK)
-			goto cleanup;
-
-		e1 = exp_fn(p, e1, e2, err);
-		e2 = NULL;
-		if (err->type != SUBTILIS_ERROR_OK)
-			goto cleanup;
-	}
-
-	return e1;
-
-cleanup:
-
-	subtilis_exp_delete(e2);
-	subtilis_exp_delete(e1);
-	return NULL;
-}
-
-static subtilis_exp_t *prv_priority6(subtilis_parser_t *p, subtilis_token_t *t,
-				     subtilis_error_t *err)
-{
-	subtilis_exp_t *e1 = NULL;
-	subtilis_exp_t *e2 = NULL;
-
-	e1 = prv_priority5(p, t, err);
-	if (err->type != SUBTILIS_ERROR_OK)
-		goto cleanup;
-	while (t->type == SUBTILIS_TOKEN_KEYWORD &&
-	       t->tok.keyword.type == SUBTILIS_KEYWORD_AND) {
-		subtilis_lexer_get(p->l, t, err);
-		if (err->type != SUBTILIS_ERROR_OK)
-			goto cleanup;
-		e2 = prv_priority5(p, t, err);
-		if (err->type != SUBTILIS_ERROR_OK)
-			goto cleanup;
-		e1 = subtilis_type_if_and(p, e1, e2, err);
-		e2 = NULL;
-		if (err->type != SUBTILIS_ERROR_OK)
-			goto cleanup;
-	}
-
-	return e1;
-
-cleanup:
-
-	subtilis_exp_delete(e2);
-	subtilis_exp_delete(e1);
-	return NULL;
-}
-
-static subtilis_exp_t *prv_priority7(subtilis_parser_t *p, subtilis_token_t *t,
-				     subtilis_error_t *err)
-{
-	subtilis_exp_t *e1 = NULL;
-	subtilis_exp_t *e2 = NULL;
-	subtilis_exp_fn_t exp_fn;
-
-	e1 = prv_priority6(p, t, err);
-	if (err->type != SUBTILIS_ERROR_OK)
-		goto cleanup;
-	while (t->type == SUBTILIS_TOKEN_KEYWORD) {
-		if (t->tok.keyword.type == SUBTILIS_KEYWORD_EOR)
-			exp_fn = subtilis_type_if_eor;
-		else if (t->tok.keyword.type == SUBTILIS_KEYWORD_OR)
-			exp_fn = subtilis_type_if_or;
-		else
-			break;
-		subtilis_lexer_get(p->l, t, err);
-		if (err->type != SUBTILIS_ERROR_OK)
-			goto cleanup;
-		e2 = prv_priority6(p, t, err);
-		if (err->type != SUBTILIS_ERROR_OK)
-			goto cleanup;
-		e1 = exp_fn(p, e1, e2, err);
-		e2 = NULL;
-		if (err->type != SUBTILIS_ERROR_OK)
-			goto cleanup;
-	}
-
-	return e1;
-
-cleanup:
-
-	subtilis_exp_delete(e2);
-	subtilis_exp_delete(e1);
-	return NULL;
-}
-
-static subtilis_exp_t *prv_expression(subtilis_parser_t *p, subtilis_token_t *t,
-				      subtilis_error_t *err)
-{
-	subtilis_lexer_get(p->l, t, err);
-	if (err->type != SUBTILIS_ERROR_OK)
-		return NULL;
-	return prv_priority7(p, t, err);
-}
-
 static subtilis_exp_t *prv_numeric_expression(subtilis_parser_t *p,
 					      subtilis_token_t *t,
 					      subtilis_error_t *err)
@@ -1366,7 +144,7 @@ static subtilis_exp_t *prv_numeric_expression(subtilis_parser_t *p,
 	subtilis_lexer_get(p->l, t, err);
 	if (err->type != SUBTILIS_ERROR_OK)
 		return NULL;
-	e = prv_priority7(p, t, err);
+	e = subtilis_parser_priority7(p, t, err);
 	if (err->type != SUBTILIS_ERROR_OK)
 		return NULL;
 	if ((e->type.type != SUBTILIS_TYPE_CONST_INTEGER) &&
@@ -1378,32 +156,6 @@ static subtilis_exp_t *prv_numeric_expression(subtilis_parser_t *p,
 	}
 
 	return e;
-}
-
-static subtilis_exp_t *prv_int_var_expression(subtilis_parser_t *p,
-					      subtilis_token_t *t,
-					      subtilis_error_t *err)
-{
-	subtilis_exp_t *e = NULL;
-
-	e = prv_priority7(p, t, err);
-	if (err->type != SUBTILIS_ERROR_OK)
-		return NULL;
-
-	e = subtilis_type_if_to_int(p, e, err);
-	if (err->type != SUBTILIS_ERROR_OK)
-		goto cleanup;
-
-	e = subtilis_type_if_exp_to_var(p, e, err);
-	if (err->type != SUBTILIS_ERROR_OK)
-		goto cleanup;
-
-	return e;
-
-cleanup:
-
-	subtilis_exp_delete(e);
-	return NULL;
 }
 
 typedef enum {
@@ -1446,68 +198,6 @@ cleanup:
 	return NULL;
 }
 
-/*
- * Returns 0 and no error if more than max args are read.  Caller is
- * expected to free expressions even in case of error.
- */
-
-static size_t prv_var_bracketed_int_args_have_b(subtilis_parser_t *p,
-						subtilis_token_t *t,
-						subtilis_exp_t **e, size_t max,
-						subtilis_error_t *err)
-{
-	size_t i;
-	const char *tbuf;
-
-	subtilis_lexer_get(p->l, t, err);
-	if (err->type != SUBTILIS_ERROR_OK)
-		return 0;
-
-	e[0] = prv_priority7(p, t, err);
-	if (err->type != SUBTILIS_ERROR_OK)
-		return 0;
-
-	e[0] = subtilis_type_if_to_int(p, e[0], err);
-	if (err->type != SUBTILIS_ERROR_OK)
-		return 0;
-
-	for (i = 1; i < max; i++) {
-		tbuf = subtilis_token_get_text(t);
-		if (t->type != SUBTILIS_TOKEN_OPERATOR) {
-			subtilis_error_set_expected(
-			    err, ",", tbuf, p->l->stream->name, p->l->line);
-			return 0;
-		}
-
-		if (!strcmp(tbuf, ")"))
-			return i;
-
-		if (strcmp(tbuf, ",")) {
-			subtilis_error_set_expected(
-			    err, ",", tbuf, p->l->stream->name, p->l->line);
-			return 0;
-		}
-
-		subtilis_lexer_get(p->l, t, err);
-		if (err->type != SUBTILIS_ERROR_OK)
-			return 0;
-
-		e[i] = prv_priority7(p, t, err);
-		if (err->type != SUBTILIS_ERROR_OK)
-			return 0;
-
-		e[i] = subtilis_type_if_to_int(p, e[i], err);
-		if (err->type != SUBTILIS_ERROR_OK)
-			return 0;
-	}
-
-	/*
-	 * Too many integers in the list
-	 */
-
-	return 0;
-}
-
 static void prv_assign_array(subtilis_parser_t *p, subtilis_token_t *t,
 			     const subtilis_symbol_t *s, size_t mem_reg,
 			     const char *var_name, subtilis_error_t *err)
@@ -1519,8 +209,8 @@ static void prv_assign_array(subtilis_parser_t *p, subtilis_token_t *t,
 	subtilis_assign_type_t at;
 	size_t dims = 0;
 
-	dims = prv_var_bracketed_int_args_have_b(p, t, indices,
-						 SUBTILIS_MAX_DIMENSIONS, err);
+	dims = subtilis_var_bracketed_int_args_have_b(
+	    p, t, indices, SUBTILIS_MAX_DIMENSIONS, err);
 	if (err->type != SUBTILIS_ERROR_OK)
 		return;
 
@@ -1547,7 +237,7 @@ static void prv_assign_array(subtilis_parser_t *p, subtilis_token_t *t,
 		goto cleanup;
 	}
 
-	e = prv_expression(p, t, err);
+	e = subtilis_parser_expression(p, t, err);
 	if (err->type != SUBTILIS_ERROR_OK)
 		goto cleanup;
 
@@ -1678,7 +368,7 @@ static void prv_assignment(subtilis_parser_t *p, subtilis_token_t *t,
 		goto cleanup;
 	}
 
-	e = prv_expression(p, t, err);
+	e = subtilis_parser_expression(p, t, err);
 	if (err->type != SUBTILIS_ERROR_OK)
 		goto cleanup;
 
@@ -1708,157 +398,6 @@ cleanup:
 	free(var_name);
 }
 
-/*
- * Returns 0 and no error if more than max args are read.  Caller is
- * expected to free expressions even in case of error.
- */
-
-static size_t prv_var_bracketed_int_args(subtilis_parser_t *p,
-					 subtilis_token_t *t,
-					 subtilis_exp_t **e, size_t max,
-					 subtilis_error_t *err)
-{
-	const char *tbuf;
-
-	subtilis_lexer_get(p->l, t, err);
-	if (err->type != SUBTILIS_ERROR_OK)
-		return 0;
-
-	tbuf = subtilis_token_get_text(t);
-	if ((t->type != SUBTILIS_TOKEN_OPERATOR) || strcmp(tbuf, "(")) {
-		subtilis_error_set_exp_expected(err, "( ", p->l->stream->name,
-						p->l->line);
-		return 0;
-	}
-
-	return prv_var_bracketed_int_args_have_b(p, t, e, max, err);
-}
-
-static void prv_deallocate_arrays(subtilis_parser_t *p,
-				  subtilis_ir_operand_t load_reg,
-				  subtilis_sizet_vector_t *free_list,
-				  subtilis_error_t *err)
-{
-	size_t i;
-
-	for (i = 0; i < free_list->len; i++) {
-		subtlis_array_type_deallocate(p, free_list->vals[i], load_reg,
-					      err);
-		if (err->type != SUBTILIS_ERROR_OK)
-			return;
-	}
-}
-
-static void prv_allocate_array(subtilis_parser_t *p, const char *var_name,
-			       subtilis_type_t *type, size_t loc,
-			       subtilis_exp_t **e,
-			       subtilis_ir_operand_t store_reg,
-			       subtilis_error_t *err)
-{
-	subtilis_sizet_vector_t *free_list;
-
-	subtlis_array_type_allocate(p, var_name, type, loc, e, store_reg, err);
-	if (err->type != SUBTILIS_ERROR_OK)
-		return;
-
-	if (store_reg.reg == SUBTILIS_IR_REG_LOCAL)
-		free_list = (p->current == p->main) ? &p->main_free_list
-						    : &p->local_free_list;
-	else
-		free_list = &p->free_list;
-
-	subtilis_sizet_vector_append(free_list, loc, err);
-}
-
-static void prv_create_array(subtilis_parser_t *p, subtilis_token_t *t,
-			     bool local, subtilis_error_t *err)
-{
-	const subtilis_symbol_t *s;
-	const char *tbuf;
-	subtilis_exp_t *e[SUBTILIS_MAX_DIMENSIONS];
-	subtilis_type_t element_type;
-	subtilis_type_t type;
-	size_t i;
-	subtilis_ir_operand_t local_global;
-	size_t dims = 0;
-	char *var_name = NULL;
-
-	local_global.reg =
-	    local ? SUBTILIS_IR_REG_LOCAL : SUBTILIS_IR_REG_GLOBAL;
-
-	do {
-		subtilis_lexer_get(p->l, t, err);
-		if (err->type != SUBTILIS_ERROR_OK)
-			return;
-		if (t->type != SUBTILIS_TOKEN_IDENTIFIER) {
-			tbuf = subtilis_token_get_text(t);
-			subtilis_error_set_id_expected(
-			    err, tbuf, p->l->stream->name, p->l->line);
-			return;
-		}
-
-		tbuf = subtilis_token_get_text(t);
-		if (subtilis_symbol_table_lookup(p->local_st, tbuf) ||
-		    (!local && subtilis_symbol_table_lookup(p->st, tbuf))) {
-			subtilis_error_set_already_defined(
-			    err, tbuf, p->l->stream->name, p->l->line);
-			return;
-		}
-
-		var_name = malloc(strlen(tbuf) + 1);
-		if (!var_name) {
-			subtilis_error_set_oom(err);
-			return;
-		}
-		strcpy(var_name, tbuf);
-		element_type = t->tok.id_type;
-
-		dims = prv_var_bracketed_int_args(p, t, e,
-						  SUBTILIS_MAX_DIMENSIONS, err);
-		if (err->type != SUBTILIS_ERROR_OK)
-			goto cleanup;
-
-		if (dims == 0) {
-			subtilis_error_too_many_dims(
-			    err, var_name, p->l->stream->name, p->l->line);
-			goto cleanup;
-		}
-
-		subtilis_array_type_init(p, &element_type, &type, &e[0], dims,
-					 err);
-		if (err->type != SUBTILIS_ERROR_OK)
-			goto cleanup;
-
-		s = subtilis_symbol_table_insert(local ? p->local_st : p->st,
-						 var_name, &type, err);
-		if (err->type != SUBTILIS_ERROR_OK)
-			goto cleanup;
-
-		prv_allocate_array(p, var_name, &type, s->loc, &e[0],
-				   local_global, err);
-		if (err->type != SUBTILIS_ERROR_OK)
-			goto cleanup;
-
-		for (i = 0; i < dims; i++)
-			subtilis_exp_delete(e[i]);
-		dims = 0;
-
-		free(var_name);
-		var_name = NULL;
-
-		subtilis_lexer_get(p->l, t, err);
-		if (err->type != SUBTILIS_ERROR_OK)
-			goto cleanup;
-		tbuf = subtilis_token_get_text(t);
-	} while (t->type == SUBTILIS_TOKEN_OPERATOR && !strcmp(tbuf, ","));
-
-cleanup:
-	for (i = 0; i < dims; i++)
-		subtilis_exp_delete(e[i]);
-
-	free(var_name);
-}
-
 static void prv_dim(subtilis_parser_t *p, subtilis_token_t *t,
 		    subtilis_error_t *err)
 {
@@ -1867,7 +406,7 @@ static void prv_dim(subtilis_parser_t *p, subtilis_token_t *t,
 		return;
 	}
 
-	prv_create_array(p, t, false, err);
+	subtilis_parser_create_array(p, t, false, err);
 }
 
 static void prv_let(subtilis_parser_t *p, subtilis_token_t *t,
@@ -1888,82 +427,6 @@ static void prv_let(subtilis_parser_t *p, subtilis_token_t *t,
 	prv_assignment(p, t, err);
 }
 
-static void prv_parse_locals(subtilis_parser_t *p, subtilis_token_t *t,
-			     subtilis_error_t *err)
-{
-	const char *tbuf;
-	subtilis_exp_t *e;
-	subtilis_type_t type;
-	char *var_name = NULL;
-
-	while ((t->type == SUBTILIS_TOKEN_KEYWORD) &&
-	       (t->tok.keyword.type == SUBTILIS_KEYWORD_LOCAL)) {
-		subtilis_lexer_get(p->l, t, err);
-		if (err->type != SUBTILIS_ERROR_OK)
-			return;
-
-		tbuf = subtilis_token_get_text(t);
-		if ((t->type == SUBTILIS_TOKEN_KEYWORD) &&
-		    (t->tok.keyword.type == SUBTILIS_KEYWORD_DIM)) {
-			prv_create_array(p, t, true, err);
-			if (err->type != SUBTILIS_ERROR_OK)
-				return;
-			continue;
-		} else if (t->type != SUBTILIS_TOKEN_IDENTIFIER) {
-			subtilis_error_set_id_expected(
-			    err, tbuf, p->l->stream->name, p->l->line);
-			return;
-		}
-
-		if (subtilis_symbol_table_lookup(p->local_st, tbuf)) {
-			subtilis_error_set_already_defined(
-			    err, tbuf, p->l->stream->name, p->l->line);
-			return;
-		}
-
-		var_name = malloc(strlen(tbuf) + 1);
-		if (!var_name) {
-			subtilis_error_set_oom(err);
-			return;
-		}
-		strcpy(var_name, tbuf);
-		type = t->tok.id_type;
-
-		subtilis_lexer_get(p->l, t, err);
-		if (err->type != SUBTILIS_ERROR_OK)
-			goto cleanup;
-
-		tbuf = subtilis_token_get_text(t);
-		if (!((t->type == SUBTILIS_TOKEN_OPERATOR) &&
-		      !strcmp(tbuf, "="))) {
-			e = subtilis_type_if_zero(p, &type, err);
-			if (err->type != SUBTILIS_ERROR_OK)
-				goto cleanup;
-		} else {
-			e = prv_expression(p, t, err);
-			if (err->type != SUBTILIS_ERROR_OK)
-				goto cleanup;
-			e = subtilis_type_if_exp_to_var(p, e, err);
-			if (err->type != SUBTILIS_ERROR_OK)
-				goto cleanup;
-			e = subtilis_exp_coerce_type(p, e, &type, err);
-			if (err->type != SUBTILIS_ERROR_OK)
-				goto cleanup;
-		}
-		(void)subtilis_symbol_table_insert_reg(
-		    p->local_st, var_name, &type, e->exp.ir_op.reg, err);
-		subtilis_exp_delete(e);
-		free(var_name);
-		var_name = NULL;
-		if (err->type != SUBTILIS_ERROR_OK)
-			return;
-	}
-
-cleanup:
-
-	free(var_name);
-}
-
 static void prv_mode(subtilis_parser_t *p, subtilis_token_t *t,
 		     subtilis_error_t *err)
 {
@@ -1973,7 +436,7 @@ static void prv_mode(subtilis_parser_t *p, subtilis_token_t *t,
 	if (err->type != SUBTILIS_ERROR_OK)
 		return;
 
-	e = prv_int_var_expression(p, t, err);
+	e = subtilis_parser_int_var_expression(p, t, err);
 	if (err->type != SUBTILIS_ERROR_OK)
 		return;
 
@@ -1989,107 +452,6 @@ cleanup:
 	subtilis_exp_delete(e);
 }
 
-static void prv_statement_int_args(subtilis_parser_t *p, subtilis_token_t *t,
-				   subtilis_exp_t **e, size_t expected,
-				   subtilis_error_t *err)
-{
-	size_t i;
-	const char *tbuf;
-
-	e[0] = prv_int_var_expression(p, t, err);
-	if (err->type != SUBTILIS_ERROR_OK)
-		return;
-
-	for (i = 1; i < expected; i++) {
-		tbuf = subtilis_token_get_text(t);
-		if (t->type != SUBTILIS_TOKEN_OPERATOR) {
-			subtilis_error_set_expected(
-			    err, ",", tbuf, p->l->stream->name, p->l->line);
-			return;
-		}
-
-		if (strcmp(tbuf, ",")) {
-			subtilis_error_set_expected(
-			    err, ",", tbuf, p->l->stream->name, p->l->line);
-			return;
-		}
-
-		subtilis_lexer_get(p->l, t, err);
-		if (err->type != SUBTILIS_ERROR_OK)
-			return;
-
-		e[i] = prv_int_var_expression(p, t, err);
-		if (err->type != SUBTILIS_ERROR_OK)
-			return;
-	}
-}
-
-static void prv_bracketed_int_args(subtilis_parser_t *p, subtilis_token_t *t,
-				   subtilis_exp_t **e, size_t expected,
-				   subtilis_error_t *err)
-{
-	const char *tbuf;
-
-	subtilis_lexer_get(p->l, t, err);
-	if (err->type != SUBTILIS_ERROR_OK)
-		return;
-
-	tbuf = subtilis_token_get_text(t);
-	if (strcmp(tbuf, "(")) {
-		subtilis_error_set_exp_expected(err, "( ", p->l->stream->name,
-						p->l->line);
-		return;
-	}
-
-	subtilis_lexer_get(p->l, t, err);
-	if (err->type != SUBTILIS_ERROR_OK)
-		return;
-
-	prv_statement_int_args(p, t, e, expected, err);
-	if (err->type != SUBTILIS_ERROR_OK)
-		return;
-
-	tbuf = subtilis_token_get_text(t);
-	if (strcmp(tbuf, ")")) {
-		subtilis_error_set_exp_expected(err, ") ", p->l->stream->name,
-						p->l->line);
-		return;
-	}
-
-	subtilis_lexer_get(p->l, t, err);
-}
-
-static subtilis_exp_t *prv_bracketed_2_int_args(subtilis_parser_t *p,
-						subtilis_token_t *t,
-						subtilis_op_instr_type_t itype,
-						subtilis_error_t *err)
-{
-	subtilis_exp_t *e[2];
-	size_t reg;
-	size_t i;
-	subtilis_exp_t *retval = NULL;
-
-	memset(&e, 0, sizeof(e));
-
-	prv_bracketed_int_args(p, t, e, 2, err);
-	if (err->type != SUBTILIS_ERROR_OK)
-		return NULL;
-
-	reg = subtilis_ir_section_add_instr(p->current, itype, e[0]->exp.ir_op,
-					    e[1]->exp.ir_op, err);
-	if (err->type != SUBTILIS_ERROR_OK)
-		goto cleanup;
-
-	retval = subtilis_exp_new_int32_var(reg, err);
-
-cleanup:
-
-	for (i = 0; i < 2; i++)
-		subtilis_exp_delete(e[i]);
-
-	return retval;
-}
-
 static void prv_plot(subtilis_parser_t *p, subtilis_token_t *t,
 		     subtilis_error_t *err)
 {
@@ -2102,7 +464,7 @@ static void prv_plot(subtilis_parser_t *p, subtilis_token_t *t,
 	if (err->type != SUBTILIS_ERROR_OK)
 		return;
 
-	prv_statement_int_args(p, t, e, 3, err);
+	subtilis_parser_statement_int_args(p, t, e, 3, err);
 	if (err->type != SUBTILIS_ERROR_OK)
 		goto cleanup;
 
@@ -2127,7 +489,7 @@ static void prv_simple_plot(subtilis_parser_t *p, subtilis_token_t *t,
 
 	memset(&e, 0, sizeof(e));
 
-	prv_statement_int_args(p, t, &e[1], 2, err);
+	subtilis_parser_statement_int_args(p, t, &e[1], 2, err);
 	if (err->type != SUBTILIS_ERROR_OK)
 		goto cleanup;
 
@@ -2266,7 +628,7 @@ static void prv_circle(subtilis_parser_t *p, subtilis_token_t *t,
 	if (err->type != SUBTILIS_ERROR_OK)
 		goto cleanup;
 
-	e[1] = prv_int_var_expression(p, t, err);
+	e[1] = subtilis_parser_int_var_expression(p, t, err);
 	if (err->type != SUBTILIS_ERROR_OK)
 		goto cleanup;
 
@@ -2313,7 +675,7 @@ static void prv_rectangle_outline(subtilis_parser_t *p, subtilis_token_t *t,
 		goto cleanup;
 	;
 
-	prv_statement_int_args(p, t, &e[2], 2, err);
+	subtilis_parser_statement_int_args(p, t, &e[2], 2, err);
 	if (err->type != SUBTILIS_ERROR_OK)
 		goto cleanup;
 
@@ -2451,7 +813,7 @@ static void prv_gcol(subtilis_parser_t *p, subtilis_token_t *t,
 	if (err->type != SUBTILIS_ERROR_OK)
 		return;
 
-	prv_statement_int_args(p, t, e, 2, err);
+	subtilis_parser_statement_int_args(p, t, e, 2, err);
 	if (err->type != SUBTILIS_ERROR_OK)
 		goto cleanup;
 
@@ -2482,7 +844,7 @@ static void prv_origin(subtilis_parser_t *p, subtilis_token_t *t,
 	if (err->type != SUBTILIS_ERROR_OK)
 		return;
 
-	prv_statement_int_args(p, t, e, 2, err);
+	subtilis_parser_statement_int_args(p, t, e, 2, err);
 	if (err->type != SUBTILIS_ERROR_OK)
 		goto cleanup;
 
@@ -2701,7 +1063,7 @@ static void prv_vdu_brackets(subtilis_parser_t *p, subtilis_token_t *t,
 		return;
 
 	do {
-		e = prv_priority7(p, t, err);
+		e = subtilis_parser_priority7(p, t, err);
 		if (err->type != SUBTILIS_ERROR_OK)
 			return;
 
@@ -2798,7 +1160,7 @@ static void prv_vdu(subtilis_parser_t *p, subtilis_token_t *t,
 	}
 
 	do {
-		e = prv_priority7(p, t, err);
+		e = subtilis_parser_priority7(p, t, err);
 		if (err->type != SUBTILIS_ERROR_OK)
 			return;
 
@@ -2859,8 +1221,9 @@ static void prv_print_fp(subtilis_parser_t *p, size_t reg,
 			return;
 	}
 
-	(void)prv_call_1_arg_fn(p, "_print_fp", reg, SUBTILIS_IR_REG_TYPE_REAL,
-				&subtilis_type_void, err);
+	(void)subtilis_parser_call_1_arg_fn(p, "_print_fp", reg,
+					    SUBTILIS_IR_REG_TYPE_REAL,
+					    &subtilis_type_void, err);
 }
 
 static void prv_print(subtilis_parser_t *p, subtilis_token_t *t,
@@ -2868,7 +1231,7 @@ static void prv_print(subtilis_parser_t *p, subtilis_token_t *t,
 {
 	subtilis_exp_t *e = NULL;
 
-	e = prv_expression(p, t, err);
+	e = subtilis_parser_expression(p, t, err);
 	if (err->type != SUBTILIS_ERROR_OK)
 		return;
 
@@ -2909,45 +1272,6 @@ cleanup:
 	subtilis_exp_delete(e);
 }
 
-static void prv_return(subtilis_parser_t *p, subtilis_token_t *t,
-		       subtilis_error_t *err)
-{
-	subtilis_exp_t *e;
-	subtilis_type_t fn_type;
-	subtilis_ir_operand_t end_label;
-
-	if (p->current == p->main) {
-		subtilis_error_set_return_in_main(err, p->l->stream->name,
-						  p->l->line);
-		return;
-	}
-
-	fn_type = p->current->type->return_type;
-	if (fn_type.type == SUBTILIS_TYPE_VOID) {
-		subtilis_error_set_return_in_proc(err, p->l->stream->name,
-						  p->l->line);
-		return;
-	}
-
-	e = prv_expression(p, t, err);
-	if (err->type != SUBTILIS_ERROR_OK)
-		return;
-
-	/* Ownership of e is passed to add_fn_ret */
-
-	prv_set_fn_retval(p, e, &fn_type, err);
-	if (err->type != SUBTILIS_ERROR_OK)
-		return;
-
-	end_label.label = p->current->end_label;
-	subtilis_ir_section_add_instr_no_reg(p->current, SUBTILIS_OP_INSTR_JMP,
-					     end_label, err);
-	if (err->type != SUBTILIS_ERROR_OK)
-		return;
-
-	p->current->endproc = true;
-}
-
 static subtilis_keyword_type_t
 prv_if_compund(subtilis_parser_t *p, subtilis_token_t *t, subtilis_error_t *err)
 {
@@ -2975,7 +1299,7 @@ prv_if_compund(subtilis_parser_t *p, subtilis_token_t *t, subtilis_error_t *err)
 			continue;
 		} else if ((t->type == SUBTILIS_TOKEN_OPERATOR) &&
 			   !strcmp(tbuf, "<-")) {
-			prv_return(p, t, err);
+			subtilis_parser_return(p, t, err);
 			if (err->type != SUBTILIS_ERROR_OK)
 				return key_type;
 			continue;
@@ -3019,8 +1343,8 @@ prv_if_compund(subtilis_parser_t *p, subtilis_token_t *t, subtilis_error_t *err)
 	return key_type;
 }
 
-static void prv_statement(subtilis_parser_t *p, subtilis_token_t *t,
-			  subtilis_error_t *err)
+void subtilis_parser_statement(subtilis_parser_t *p, subtilis_token_t *t,
+			       subtilis_error_t *err)
 {
 	subtilis_keyword_fn fn;
 	const char *tbuf;
@@ -3040,7 +1364,7 @@ static void prv_statement(subtilis_parser_t *p, subtilis_token_t *t,
 		return;
 	} else if ((t->type == SUBTILIS_TOKEN_OPERATOR) &&
 		   !strcmp(tbuf, "<-")) {
-		prv_return(p, t, err);
+		subtilis_parser_return(p, t, err);
 		return;
 	} else if (t->type != SUBTILIS_TOKEN_KEYWORD) {
 		subtilis_error_set_keyword_expected(
@@ -3059,66 +1383,6 @@ static void prv_statement(subtilis_parser_t *p, subtilis_token_t *t,
 	fn(p, t, err);
 }
 
-static void prv_compound(subtilis_parser_t *p, subtilis_token_t *t,
-			 subtilis_keyword_type_t end_key, subtilis_error_t *err)
-{
-	unsigned int start;
-
-	p->level++;
-	start = p->l->line;
-	while (t->type != SUBTILIS_TOKEN_EOF) {
-		if ((t->type == SUBTILIS_TOKEN_KEYWORD) &&
-		    (t->tok.keyword.type == end_key)) {
-			if ((end_key != SUBTILIS_KEYWORD_ENDPROC) ||
-			    (p->level == 1))
-				break;
-		}
-		prv_statement(p, t, err);
-		if (err->type != SUBTILIS_ERROR_OK)
-			return;
-	}
-
-	if (t->type == SUBTILIS_TOKEN_EOF)
-		subtilis_error_set_compund_not_term(err, p->l->stream->name,
-						    start);
-
-	if ((end_key != SUBTILIS_KEYWORD_NEXT) &&
-	    (end_key != SUBTILIS_KEYWORD_UNTIL) &&
-	    (end_key != SUBTILIS_KEYWORD_ENDPROC))
-		p->current->endproc = false;
-	p->level--;
-	p->current->handler_list =
-	    subtilis_handler_list_truncate(p->current->handler_list, p->level);
-}
-
-static void prv_fn_compound(subtilis_parser_t *p, subtilis_token_t *t,
-			    subtilis_error_t *err)
-{
-	unsigned int start;
-	const char *tbuf;
-
-	p->level++;
-	start = p->l->line;
-	while (t->type != SUBTILIS_TOKEN_EOF) {
-		tbuf = subtilis_token_get_text(t);
-		if ((t->type == SUBTILIS_TOKEN_OPERATOR) &&
-		    !strcmp(tbuf, "<-") && (p->level == 1))
-			break;
-
-		prv_statement(p, t, err);
-		if (err->type != SUBTILIS_ERROR_OK)
-			return;
-	}
-
-	if (t->type == SUBTILIS_TOKEN_EOF)
-		subtilis_error_set_compund_not_term(err, p->l->stream->name,
-						    start);
-	//	p->current->endproc = false;
-	p->level--;
-	p->current->handler_list =
-	    subtilis_handler_list_truncate(p->current->handler_list, p->level);
-}
-
 static subtilis_exp_t *prv_conditional_exp(subtilis_parser_t *p,
 					   subtilis_token_t *t,
 					   subtilis_ir_operand_t *cond,
@@ -3127,7 +1391,7 @@ static subtilis_exp_t *prv_conditional_exp(subtilis_parser_t *p,
 	subtilis_exp_t *e;
 	size_t reg;
 
-	e = prv_expression(p, t, err);
+	e = subtilis_parser_expression(p, t, err);
 	if (err->type != SUBTILIS_ERROR_OK)
 		return NULL;
 
@@ -3206,7 +1470,7 @@ static void prv_if(subtilis_parser_t *p, subtilis_token_t *t,
 		subtilis_lexer_get(p->l, t, err);
 		if (err->type != SUBTILIS_ERROR_OK)
 			return;
-		prv_compound(p, t, SUBTILIS_KEYWORD_ENDIF, err);
+		subtilis_parser_compound(p, t, SUBTILIS_KEYWORD_ENDIF, err);
 		if (err->type != SUBTILIS_ERROR_OK)
 			goto cleanup;
 		subtilis_ir_section_add_label(p->current, end_label.reg, err);
@@ -3224,21 +1488,6 @@ static void prv_if(subtilis_parser_t *p, subtilis_token_t *t,
 
 cleanup:
 	subtilis_exp_delete(e);
-}
-
-static void prv_handle_escape(subtilis_parser_t *p, subtilis_error_t *err)
-{
-	/* Let's not test for escape in an error handler. */
-
-	if (!p->handle_escapes || p->current->in_error_handler)
-		return;
-
-	subtilis_ir_section_add_instr_no_arg(p->current,
-					     SUBTILIS_OP_INSTR_TESTESC, err);
-	if (err->type != SUBTILIS_ERROR_OK)
-		return;
-
-	subtilis_exp_handle_errors(p, err);
 }
 
 static void prv_while(subtilis_parser_t *p, subtilis_token_t *t,
@@ -3271,11 +1520,11 @@ static void prv_while(subtilis_parser_t *p, subtilis_token_t *t,
 	if (err->type != SUBTILIS_ERROR_OK)
 		goto cleanup;
 
-	prv_handle_escape(p, err);
+	subtilis_parser_handle_escape(p, err);
 	if (err->type != SUBTILIS_ERROR_OK)
 		goto cleanup;
 
-	prv_compound(p, t, SUBTILIS_KEYWORD_ENDWHILE, err);
+	subtilis_parser_compound(p, t, SUBTILIS_KEYWORD_ENDWHILE, err);
 	if (err->type != SUBTILIS_ERROR_OK)
 		goto cleanup;
 
@@ -3313,11 +1562,11 @@ static void prv_repeat(subtilis_parser_t *p, subtilis_token_t *t,
 	if (err->type != SUBTILIS_ERROR_OK)
 		goto cleanup;
 
-	prv_handle_escape(p, err);
+	subtilis_parser_handle_escape(p, err);
 	if (err->type != SUBTILIS_ERROR_OK)
 		goto cleanup;
 
-	prv_compound(p, t, SUBTILIS_KEYWORD_UNTIL, err);
+	subtilis_parser_compound(p, t, SUBTILIS_KEYWORD_UNTIL, err);
 	if (err->type != SUBTILIS_ERROR_OK)
 		goto cleanup;
 
@@ -3443,8 +1692,8 @@ static void prv_for_assign_array(subtilis_parser_t *p, const char *var_name,
 	subtilis_exp_t *e = NULL;
 	size_t dims = 0;
 
-	dims = prv_var_bracketed_int_args_have_b(p, t, indices,
-						 SUBTILIS_MAX_DIMENSIONS, err);
+	dims = subtilis_var_bracketed_int_args_have_b(
+	    p, t, indices, SUBTILIS_MAX_DIMENSIONS, err);
 	if (err->type != SUBTILIS_ERROR_OK)
 		return;
 
@@ -3526,7 +1775,7 @@ static void prv_for_assignment(subtilis_parser_t *p, subtilis_token_t *t,
 		goto cleanup;
 	}
 
-	e = prv_expression(p, t, err);
+	e = subtilis_parser_expression(p, t, err);
 	if (err->type != SUBTILIS_ERROR_OK)
 		goto cleanup;
 
@@ -3557,11 +1806,11 @@ static void prv_for_loop_start(subtilis_parser_t *p, subtilis_token_t *t,
 	if (err->type != SUBTILIS_ERROR_OK)
 		return;
 
-	prv_handle_escape(p, err);
+	subtilis_parser_handle_escape(p, err);
 	if (err->type != SUBTILIS_ERROR_OK)
 		return;
 
-	prv_compound(p, t, SUBTILIS_KEYWORD_NEXT, err);
+	subtilis_parser_compound(p, t, SUBTILIS_KEYWORD_NEXT, err);
 	if (err->type != SUBTILIS_ERROR_OK)
 		return;
 
@@ -4025,7 +2274,7 @@ static void prv_onerror(subtilis_parser_t *p, subtilis_token_t *t,
 		if ((t->type == SUBTILIS_TOKEN_KEYWORD) &&
 		    (t->tok.keyword.type == SUBTILIS_KEYWORD_ENDERROR))
 			break;
-		prv_statement(p, t, err);
+		subtilis_parser_statement(p, t, err);
 		if (err->type != SUBTILIS_ERROR_OK)
 			goto cleanup;
 	}
@@ -4077,7 +2326,7 @@ static void prv_error(subtilis_parser_t *p, subtilis_token_t *t,
 	if (err->type != SUBTILIS_ERROR_OK)
 		return;
 
-	e = prv_priority7(p, t, err);
+	e = subtilis_parser_priority7(p, t, err);
 	if (err->type != SUBTILIS_ERROR_OK)
 		return;
 
@@ -4094,522 +2343,10 @@ static void prv_error(subtilis_parser_t *p, subtilis_token_t *t,
 	p->current->endproc = true;
 }
 
-static subtilis_type_t *prv_def_parameters(subtilis_parser_t *p,
-					   subtilis_token_t *t,
-					   size_t *num_parameters,
-					   subtilis_error_t *err)
-{
-	const char *tbuf;
-	subtilis_type_t *new_params;
-	size_t new_max;
-	size_t max_params = 0;
-	size_t num_params = 0;
-	size_t num_iparams = 0;
-	size_t num_fparams = 0;
-	subtilis_type_t *params = NULL;
-	size_t reg_num;
-
-	subtilis_lexer_get(p->l, t, err);
-	if (err->type != SUBTILIS_ERROR_OK)
-		return NULL;
-
-	tbuf = subtilis_token_get_text(t);
-	if ((t->type == SUBTILIS_TOKEN_OPERATOR) && !strcmp(tbuf, ")"))
-		goto read_next;
-
-	for (;;) {
-		if (t->type != SUBTILIS_TOKEN_IDENTIFIER) {
-			subtilis_error_set_id_expected(
-			    err, tbuf, p->l->stream->name, p->l->line);
-			goto on_error;
-		}
-
-		/* TODO: maybe these checks should go in symbol table insert */
-
-		if (subtilis_symbol_table_lookup(p->local_st, tbuf)) {
-			subtilis_error_set_already_defined(
-			    err, tbuf, p->l->stream->name, p->l->line);
-			goto on_error;
-		}
-
-		switch (t->tok.id_type.type) {
-		case SUBTILIS_TYPE_INTEGER:
-			reg_num = SUBTILIS_IR_REG_TEMP_START + num_iparams;
-			num_iparams++;
-			break;
-		case SUBTILIS_TYPE_REAL:
-			reg_num = num_fparams++;
-			break;
-		default:
-			subtilis_error_set_assertion_failed(err);
-			goto on_error;
-		}
-
-		(void)subtilis_symbol_table_insert_reg(
-		    p->local_st, tbuf, &t->tok.id_type, reg_num, err);
-		if (err->type != SUBTILIS_ERROR_OK)
-			goto on_error;
-
-		if (num_params == max_params) {
-			new_max = max_params + 16;
-			new_params = realloc(params, new_max * sizeof(*params));
-			if (!new_params) {
-				subtilis_error_set_oom(err);
-				goto on_error;
-			}
-			max_params = new_max;
-			params = new_params;
-		}
-		params[num_params++] = t->tok.id_type;
-
-		subtilis_lexer_get(p->l, t, err);
-		if (err->type != SUBTILIS_ERROR_OK)
-			goto on_error;
-		tbuf = subtilis_token_get_text(t);
-
-		if (t->type != SUBTILIS_TOKEN_OPERATOR) {
-			subtilis_error_set_right_bkt_expected(
-			    err, tbuf, p->l->stream->name, p->l->line);
-			goto on_error;
-		}
-
-		if (!strcmp(tbuf, ")"))
-			goto read_next;
-
-		if (strcmp(tbuf, ",")) {
-			subtilis_error_set_right_bkt_expected(
-			    err, tbuf, p->l->stream->name, p->l->line);
-		}
-
-		subtilis_lexer_get(p->l, t, err);
-		if (err->type != SUBTILIS_ERROR_OK)
-			goto on_error;
-		tbuf = subtilis_token_get_text(t);
-	}
-
-read_next:
-
-	*num_parameters = num_params;
-	subtilis_lexer_get(p->l, t, err);
-	return params;
-
-on_error:
-
-	free(params);
-	return NULL;
-}
-
-static char *prv_proc_name_and_type(subtilis_parser_t *p, subtilis_token_t *t,
-				    subtilis_type_t *type,
-				    subtilis_error_t *err)
-{
-	const char *tbuf;
-	subtilis_type_t id_type;
-	char *name = NULL;
-
-	tbuf = subtilis_token_get_text(t);
-	if (t->type != SUBTILIS_TOKEN_KEYWORD) {
-		subtilis_error_set_expected(err, "PROC or FN", tbuf,
-					    p->l->stream->name, p->l->line);
-		return NULL;
-	}
-
-	if (t->tok.keyword.type == SUBTILIS_KEYWORD_PROC)
-		tbuf += 4;
-	else if (t->tok.keyword.type == SUBTILIS_KEYWORD_FN)
-		tbuf += 2;
-	id_type = t->tok.keyword.id_type;
-
-	name = malloc(strlen(tbuf) + 1);
-	if (!name) {
-		subtilis_error_set_oom(err);
-		return NULL;
-	}
-
-	*type = id_type;
-	strcpy(name, tbuf);
-
-	return name;
-}
-
-static void prv_add_fn_ret(subtilis_parser_t *p, const subtilis_type_t *fn_type,
-			   subtilis_error_t *err)
-{
-	subtilis_op_instr_type_t type;
-	subtilis_ir_operand_t ret_reg;
-
-	switch (fn_type->type) {
-	case SUBTILIS_TYPE_INTEGER:
-		type = SUBTILIS_OP_INSTR_RET_I32;
-		break;
-	case SUBTILIS_TYPE_REAL:
-		type = SUBTILIS_OP_INSTR_RET_REAL;
-		break;
-	default:
-		subtilis_error_set_assertion_failed(err);
-		return;
-	}
-
-	ret_reg.reg = p->current->ret_reg;
-	subtilis_ir_section_add_instr_no_reg(p->current, type, ret_reg, err);
-}
-
-static void prv_set_fn_retval(subtilis_parser_t *p, subtilis_exp_t *e,
-			      const subtilis_type_t *fn_type,
-			      subtilis_error_t *err)
-{
-	subtilis_op_instr_type_t type;
-	subtilis_ir_operand_t ret_reg;
-
-	/* Ownership of e is passed to prv_coerce_type */
-
-	e = subtilis_exp_coerce_type(p, e, fn_type, err);
-	if (err->type != SUBTILIS_ERROR_OK)
-		return;
-
-	switch (e->type.type) {
-	case SUBTILIS_TYPE_CONST_INTEGER:
-		type = SUBTILIS_OP_INSTR_MOVI_I32;
-		break;
-	case SUBTILIS_TYPE_INTEGER:
-		type = SUBTILIS_OP_INSTR_MOV;
-		break;
-	case SUBTILIS_TYPE_CONST_REAL:
-		type = SUBTILIS_OP_INSTR_MOVI_REAL;
-		break;
-	case SUBTILIS_TYPE_REAL:
-		type = SUBTILIS_OP_INSTR_MOVFP;
-		break;
-	default:
-		subtilis_error_set_assertion_failed(err);
-		goto cleanup;
-	}
-
-	ret_reg.reg = p->current->ret_reg;
-	subtilis_ir_section_add_instr_no_reg2(p->current, type, ret_reg,
-					      e->exp.ir_op, err);
-
-cleanup:
-
-	subtilis_exp_delete(e);
-}
-
-static void prv_def(subtilis_parser_t *p, subtilis_token_t *t,
-		    subtilis_error_t *err)
-{
-	const char *tbuf;
-	subtilis_type_section_t *stype;
-	subtilis_type_t fn_type;
-	subtilis_exp_t *e;
-	subtilis_ir_operand_t var_reg;
-	size_t num_params = 0;
-	subtilis_type_t *params = NULL;
-	subtilis_symbol_table_t *local_st = NULL;
-	char *name = NULL;
-
-	if (p->current != p->main) {
-		subtilis_error_set_nested_procedure(err, p->l->stream->name,
-						    p->l->line);
-		return;
-	}
-
-	var_reg.reg = SUBTILIS_IR_REG_LOCAL;
-	subtilis_lexer_get(p->l, t, err);
-	if (err->type != SUBTILIS_ERROR_OK)
-		return;
-
-	name = prv_proc_name_and_type(p, t, &fn_type, err);
-	if (err->type != SUBTILIS_ERROR_OK)
-		return;
-
-	local_st = subtilis_symbol_table_new(err);
-	p->local_st = local_st;
-	if (err->type != SUBTILIS_ERROR_OK)
-		goto on_error;
-
-	subtilis_lexer_get(p->l, t, err);
-	if (err->type != SUBTILIS_ERROR_OK)
-		goto on_error;
-
-	tbuf = subtilis_token_get_text(t);
-	if ((t->type == SUBTILIS_TOKEN_OPERATOR) && !strcmp(tbuf, "(")) {
-		params = prv_def_parameters(p, t, &num_params, err);
-		if (err->type != SUBTILIS_ERROR_OK)
-			goto on_error;
-	}
-
-	stype = subtilis_type_section_new(&fn_type, num_params, params, err);
-	if (err->type != SUBTILIS_ERROR_OK)
-		goto on_error;
-	params = NULL;
-
-	p->current = subtilis_ir_prog_section_new(
-	    p->prog, name, p->local_st->allocated, stype, SUBTILIS_BUILTINS_MAX,
-	    p->l->stream->name, p->l->line, p->eflag_offset, p->error_offset,
-	    err);
-	if (err->type != SUBTILIS_ERROR_OK) {
-		subtilis_type_section_delete(stype);
-		goto on_error;
-	}
-
-	prv_parse_locals(p, t, err);
-	if (err->type != SUBTILIS_ERROR_OK)
-		goto on_error;
-
-	prv_handle_escape(p, err);
-	if (err->type != SUBTILIS_ERROR_OK)
-		goto on_error;
-
-	if (fn_type.type != SUBTILIS_TYPE_VOID) {
-		prv_fn_compound(p, t, err);
-		e = prv_expression(p, t, err);
-		if (err->type != SUBTILIS_ERROR_OK)
-			goto on_error;
-
-		/* Ownership of e is passed to add_fn_ret */
-
-		if (p->current->endproc) {
-			subtilis_exp_delete(e);
-		} else {
-			prv_set_fn_retval(p, e, &fn_type, err);
-			if (err->type != SUBTILIS_ERROR_OK)
-				goto on_error;
-		}
-		subtilis_ir_section_add_label(p->current, p->current->end_label,
-					      err);
-		if (err->type != SUBTILIS_ERROR_OK)
-			goto on_error;
-
-		prv_deallocate_arrays(p, var_reg, &p->local_free_list, err);
-		if (err->type != SUBTILIS_ERROR_OK)
-			goto on_error;
-
-		prv_add_fn_ret(p, &fn_type, err);
-		if (err->type != SUBTILIS_ERROR_OK)
-			goto on_error;
-	} else {
-		prv_compound(p, t, SUBTILIS_KEYWORD_ENDPROC, err);
-		if (err->type != SUBTILIS_ERROR_OK)
-			goto on_error;
-		subtilis_ir_section_add_label(p->current, p->current->end_label,
-					      err);
-		if (err->type != SUBTILIS_ERROR_OK)
-			goto on_error;
-
-		prv_deallocate_arrays(p, var_reg, &p->local_free_list, err);
-		if (err->type != SUBTILIS_ERROR_OK)
-			goto on_error;
-
-		subtilis_ir_section_add_instr_no_arg(
-		    p->current, SUBTILIS_OP_INSTR_RET, err);
-		if (err->type != SUBTILIS_ERROR_OK)
-			goto on_error;
-		subtilis_lexer_get(p->l, t, err);
-	}
-
-	if (err->type != SUBTILIS_ERROR_OK)
-		goto on_error;
-
-	subtilis_ir_merge_errors(p->current, err);
-	if (err->type != SUBTILIS_ERROR_OK)
-		goto on_error;
-
-	subtilis_array_gen_index_error_code(p, err);
-
-on_error:
-
-	free(params);
-	subtilis_symbol_table_delete(local_st);
-	free(name);
-	p->local_st = p->main_st;
-	p->current = p->main;
-	subtilis_sizet_vector_free(&p->local_free_list);
-	subtilis_sizet_vector_init(&p->local_free_list);
-}
-
-static subtilis_parser_param_t *prv_call_parameters(subtilis_parser_t *p,
-						    subtilis_token_t *t,
-						    size_t *num_parameters,
-						    subtilis_error_t *err)
-{
-	const char *tbuf;
-	subtilis_parser_param_t *new_params;
-	size_t nop;
-	size_t new_max;
-	size_t max_params = 0;
-	size_t num_params = 0;
-	subtilis_parser_param_t *params = NULL;
-	subtilis_exp_t *e = NULL;
-
-	for (;;) {
-		e = prv_expression(p, t, err);
-		if (err->type != SUBTILIS_ERROR_OK)
-			goto on_error;
-
-		if (num_params == max_params) {
-			new_max = max_params + 16;
-			new_params = realloc(params, new_max * sizeof(*params));
-			if (!new_params) {
-				subtilis_error_set_oom(err);
-				goto on_error;
-			}
-			max_params = new_max;
-			params = new_params;
-		}
-
-		tbuf = subtilis_token_get_text(t);
-
-		e = subtilis_type_if_exp_to_var(p, e, err);
-		if (err->type != SUBTILIS_ERROR_OK)
-			goto on_error;
-		params[num_params].type = e->type;
-		nop = subtilis_ir_section_add_nop(p->current, err);
-		if (err->type != SUBTILIS_ERROR_OK)
-			goto on_error;
-		params[num_params].reg = e->exp.ir_op.reg;
-		params[num_params].nop = nop;
-		num_params++;
-		subtilis_exp_delete(e);
-		e = NULL;
-
-		tbuf = subtilis_token_get_text(t);
-		if (t->type != SUBTILIS_TOKEN_OPERATOR) {
-			subtilis_error_set_right_bkt_expected(
-			    err, tbuf, p->l->stream->name, p->l->line);
-			goto on_error;
-		}
-
-		if (!strcmp(tbuf, ")"))
-			goto read_next;
-
-		if (strcmp(tbuf, ",")) {
-			subtilis_error_set_right_bkt_expected(
-			    err, tbuf, p->l->stream->name, p->l->line);
-			goto on_error;
-		}
-	}
-
-read_next:
-
-	*num_parameters = num_params;
-	subtilis_lexer_get(p->l, t, err);
-	return params;
-
-on_error:
-
-	subtilis_exp_delete(e);
-	free(params);
-	return NULL;
-}
-
-static subtilis_ir_arg_t *prv_parser_to_ir_args(subtilis_parser_param_t *params,
-						size_t num_params,
-						subtilis_error_t *err)
-{
-	size_t i;
-	subtilis_ir_arg_t *args = malloc(sizeof(*args) * num_params);
-
-	if (!args) {
-		subtilis_error_set_oom(err);
-		return NULL;
-	}
-
-	for (i = 0; i < num_params; i++) {
-		switch (params[i].type.type) {
-		case SUBTILIS_TYPE_REAL:
-			args[i].type = SUBTILIS_IR_REG_TYPE_REAL;
-			break;
-		case SUBTILIS_TYPE_INTEGER:
-			args[i].type = SUBTILIS_IR_REG_TYPE_INTEGER;
-			break;
-		default:
-			subtilis_error_set_assertion_failed(err);
-			free(args);
-			return NULL;
-		}
-		args[i].reg = params[i].reg;
-		args[i].nop = params[i].nop;
-	}
-
-	return args;
-}
-
-static subtilis_type_t *prv_parser_to_ptypes(subtilis_parser_param_t *params,
-					     size_t num_params,
-					     subtilis_error_t *err)
-{
-	size_t i;
-	subtilis_type_t *ptypes = malloc(sizeof(*ptypes) * num_params);
-
-	if (!ptypes) {
-		subtilis_error_set_oom(err);
-		return NULL;
-	}
-
-	for (i = 0; i < num_params; i++)
-		ptypes[i] = params[i].type;
-
-	return ptypes;
-}
-
-static subtilis_exp_t *prv_call(subtilis_parser_t *p, subtilis_token_t *t,
-				subtilis_error_t *err)
-{
-	const char *tbuf;
-	subtilis_type_t fn_type;
-	subtilis_type_section_t *stype = NULL;
-	subtilis_type_t *ptypes = NULL;
-	char *name = NULL;
-	subtilis_parser_param_t *params = NULL;
-	size_t num_params = 0;
-	subtilis_ir_arg_t *args = NULL;
-
-	name = prv_proc_name_and_type(p, t, &fn_type, err);
-	if (err->type != SUBTILIS_ERROR_OK)
-		return NULL;
-
-	subtilis_lexer_get(p->l, t, err);
-	if (err->type != SUBTILIS_ERROR_OK)
-		goto on_error;
-
-	tbuf = subtilis_token_get_text(t);
-	if ((t->type == SUBTILIS_TOKEN_OPERATOR) && !strcmp(tbuf, "(")) {
-		params = prv_call_parameters(p, t, &num_params, err);
-		if (err->type != SUBTILIS_ERROR_OK)
-			goto on_error;
-		args = prv_parser_to_ir_args(params, num_params, err);
-		if (err->type != SUBTILIS_ERROR_OK)
-			goto on_error;
-		ptypes = prv_parser_to_ptypes(params, num_params, err);
-		if (err->type != SUBTILIS_ERROR_OK)
-			goto on_error;
-	}
-
-	free(params);
-
-	stype = subtilis_type_section_new(&fn_type, num_params, ptypes, err);
-	if (err->type != SUBTILIS_ERROR_OK)
-		goto on_error;
-
-	/* Ownership of stypes, args and name passed to this function. */
-
-	return subtilis_exp_add_call(p, name, SUBTILIS_BUILTINS_MAX, stype,
-				     args, &fn_type, num_params, err);
-
-on_error:
-
-	free(params);
-	free(ptypes);
-	free(args);
-	free(name);
-
-	return NULL;
-}
-
 static void prv_proc(subtilis_parser_t *p, subtilis_token_t *t,
 		     subtilis_error_t *err)
 {
-	(void)prv_call(p, t, err);
+	(void)subtilis_parser_call(p, t, err);
 }
 
 static void prv_endproc(subtilis_parser_t *p, subtilis_token_t *t,
@@ -4676,12 +2413,12 @@ static void prv_root(subtilis_parser_t *p, subtilis_token_t *t,
 	if (err->type != SUBTILIS_ERROR_OK)
 		return;
 
-	prv_parse_locals(p, t, err);
+	subtilis_parser_locals(p, t, err);
 	if (err->type != SUBTILIS_ERROR_OK)
 		return;
 	p->main->locals = p->main_st->allocated;
 	while (t->type != SUBTILIS_TOKEN_EOF) {
-		prv_statement(p, t, err);
+		subtilis_parser_statement(p, t, err);
 		if (err->type != SUBTILIS_ERROR_OK)
 			return;
 	}
@@ -4691,12 +2428,12 @@ static void prv_root(subtilis_parser_t *p, subtilis_token_t *t,
 		return;
 
 	var_reg.reg = SUBTILIS_IR_REG_LOCAL;
-	prv_deallocate_arrays(p, var_reg, &p->main_free_list, err);
+	subtilis_parser_deallocate_arrays(p, var_reg, &p->main_free_list, err);
 	if (err->type != SUBTILIS_ERROR_OK)
 		return;
 
 	var_reg.reg = SUBTILIS_IR_REG_GLOBAL;
-	prv_deallocate_arrays(p, var_reg, &p->free_list, err);
+	subtilis_parser_deallocate_arrays(p, var_reg, &p->free_list, err);
 	if (err->type != SUBTILIS_ERROR_OK)
 		return;
 
@@ -4898,7 +2635,7 @@ static const subtilis_keyword_fn keyword_fns[] = {
 	NULL, /* SUBTILIS_KEYWORD_COUNT */
 	NULL, /* SUBTILIS_KEYWORD_CRUNCH */
 	NULL, /* SUBTILIS_KEYWORD_DATA */
-	prv_def, /* SUBTILIS_KEYWORD_DEF */
+	subtilis_parser_def, /* SUBTILIS_KEYWORD_DEF */
 	NULL, /* SUBTILIS_KEYWORD_DEG */
 	NULL, /* SUBTILIS_KEYWORD_DELETE */
 	prv_dim, /* SUBTILIS_KEYWORD_DIM */
